requirements.txt
PyAudio==0.2.14
sparkai==0.3.0
websocket_client==1.8.0
wxPython==4.2.2
